Designed HW514b
Designed and implemented HW514b which takes 1 fewer cycle to complete than HW514 with a slack of 0.045ns. It does this by immediatly calculating the first half of the inputs instead of waiting for the input to go through the register. It reduces the input register size since they now only store 514 bits. Also removed one of the 2 carry muxs since it was redundant